+++ tests/garage/torch/policies/test_gaussian_mlp_policy.py
@@ -36,12 +36,15 @@ class TestGaussianMLPPolicies:
         dist = policy(obs)[0]
 
         expected_mean = torch.full(
-            (act_dim, ), obs_dim * (torch.Tensor(hidden_sizes).prod().item()))
+            (act_dim, ),
+            obs_dim * (torch.Tensor(hidden_sizes).prod().item()),
+            dtype=torch.float)
         expected_variance = init_std**2
         action, prob = policy.get_action(obs)
 
         assert np.array_equal(prob['mean'], expected_mean.numpy())
-        assert dist.variance.equal(torch.full((act_dim, ), expected_variance))
+        assert dist.variance.equal(
+            torch.full((act_dim, ), expected_variance, dtype=torch.float))
         assert action.shape == (act_dim, )
 
     # yapf: disable
@@ -67,12 +70,15 @@ class TestGaussianMLPPolicies:
         dist = policy(torch.from_numpy(obs))[0]
 
         expected_mean = torch.full(
-            (act_dim, ), obs_dim * (torch.Tensor(hidden_sizes).prod().item()))
+            (act_dim, ),
+            obs_dim * (torch.Tensor(hidden_sizes).prod().item()),
+            dtype=torch.float)
         expected_variance = init_std**2
         action, prob = policy.get_action(obs)
 
         assert np.array_equal(prob['mean'], expected_mean.numpy())
-        assert dist.variance.equal(torch.full((act_dim, ), expected_variance))
+        assert dist.variance.equal(
+            torch.full((act_dim, ), expected_variance, dtype=torch.float))
         assert action.shape == (act_dim, )
 
     # yapf: disable
@@ -104,13 +110,16 @@ class TestGaussianMLPPolicies:
 
         expected_mean = torch.full([batch_size, act_dim],
                                    obs_dim *
-                                   (torch.Tensor(hidden_sizes).prod().item()))
+                                   (torch.Tensor(hidden_sizes).prod().item()),
+                                   dtype=torch.float)
         expected_variance = init_std**2
         action, prob = policy.get_actions(obs)
 
         assert np.array_equal(prob['mean'], expected_mean.numpy())
         assert dist.variance.equal(
-            torch.full((batch_size, act_dim), expected_variance))
+            torch.full((batch_size, act_dim),
+                       expected_variance,
+                       dtype=torch.float))
         assert action.shape == (batch_size, act_dim)
 
     # yapf: disable
@@ -142,13 +151,16 @@ class TestGaussianMLPPolicies:
 
         expected_mean = torch.full([batch_size, act_dim],
                                    obs_dim *
-                                   (torch.Tensor(hidden_sizes).prod().item()))
+                                   (torch.Tensor(hidden_sizes).prod().item()),
+                                   dtype=torch.float)
         expected_variance = init_std**2
         action, prob = policy.get_actions(obs)
 
         assert np.array_equal(prob['mean'], expected_mean.numpy())
         assert dist.variance.equal(
-            torch.full((batch_size, act_dim), expected_variance))
+            torch.full((batch_size, act_dim),
+                       expected_variance,
+                       dtype=torch.float))
         assert action.shape == (batch_size, act_dim)
 
     # yapf: disable
